// src/app/store/game.action.enum.ts
export enum GameActions {
  SET_QUESTIONS = '[Game] Save questions',
  SET_ACTUAL_QUESTION = '[Game] Set actual question',
  SET_GAME_SETTINGS = '[Game] Set game settings',
  NEXT_STEP = '[Game] Next step',
  NEW_GAME = '[Game] New game',
  START_OVER = '[Game] Start over',
  GET_NUMBERS_OF_QUESTIONS = '[Game] Get numbers of questions',
  GET_NUMBERS_OF_QUESTIONS_SUCCESS = '[Game] Get numbers of questions success',
  GET_NUMBERS_OF_QUESTIONS_FAILED = '[Game] Get numbers of questions failed',
  GET_PAY_DATA = '[Game] Get pay data',
  GET_PAY_DATA_SUCCESS = '[Game] Get pay data success',
  GET_PAY_DATA_FAILED = '[Game] Get pay data failed',
}